refactor: 이벤트 모달 Layout 추출 (#124)
