+++ front/src/components/calendar/EventBar/style.module.css
@@ -3,7 +3,7 @@
 .event_container:hover > *{ filter: brightness(0.9); } 
 
 .event_bar { width: 100%; height: 100%; border-radius: 3px; display: flex; align-items: center; cursor: pointer;  z-index: 3; } 
-.event_bar > em { font-size: 0.72rem; vertical-align: middle; color: #fff; text-align: left; padding:0px 10px; white-space: nowrap; overflow:hidden; text-overflow: ellipsis; } 
+.event_bar > em { width: 100%; font-size: 0.72rem; vertical-align: middle; color: #fff; text-align: left; padding:0px 10px; white-space: nowrap; overflow:hidden; text-overflow: ellipsis; } 
 
 .event_left { position: relative; border: 0.65rem solid transparent; border-left: 0px; border-right: 0.65rem solid rgb(192, 202, 51); left: 1.3px; z-index: 4; } 
 .event_right { position: relative; border: 0.65rem solid transparent; border-right: 0px; border-left: 0.65rem solid rgb(192, 202, 51); right: 1.3px; z-index: 4; }
