+++ front/src/components/modal/EventDetailModal/index.jsx
