+++ front/src/components/modal/EventDetailModal/style.module.css
